move file location from login to common. Will be using these files in different directories
