+++ app/src/main/java/com/example/tasky/di/AppModule.kt
@@ -1,11 +1,11 @@
 package com.example.tasky.di
 
 import android.app.Application
+import com.example.tasky.common.data.remote.ApiKeyInterceptor
+import com.example.tasky.common.data.remote.TaskyApi
 import com.example.tasky.common.domain.Constants
 import com.example.tasky.common.domain.EmailPatternValidator
 import com.example.tasky.common.domain.util.EmailPatternValidatorImpl
-import com.example.tasky.feature_login.data.remote.ApiKeyInterceptor
-import com.example.tasky.feature_login.data.remote.TaskyApi
 import com.example.tasky.feature_login.data.repository.UserRemoteRemoteRepositoryImpl
 import com.example.tasky.feature_login.domain.repository.UserRemoteRepository
 import dagger.Module
